Move trecho de PidProviderAPI de controller para client
